+++ examples/paraToSystemTransferMultiAssetWithFee.ts
@@ -9,7 +9,7 @@ import { GREEN, PURPLE, RESET } from './colors';
 
 /**
  * In this example we are creating a call to send 1 xcRMRK from a Moonriver (Parachain) account
- * to a Kusama Asset Hub (System Parachain) account, where the `xcmVersion` is set to 3 and a `weightLimit` is provided declaring that
+ * to a Kusama Asset Hub (System Parachain) account, where the `xcmVersion` is set to safeXcmVersion and a `weightLimit` is provided declaring that
  * it will be a weight limited tx with a custom `refTime` and `proofSize`.
  *
  * NOTE: To specify the amount of weight for the tx to use provide a `weightLimit` option containing desired values for `refTime` and `proofSize`.
@@ -30,7 +30,7 @@ const main = async () => {
 					refTime: '10000',
 					proofSize: '3000',
 				},
-				xcmVersion: 3,
+				xcmVersion: safeXcmVersion,
 				// NOTE: for xTokens `transferMultiassetWithFee` txs, paysWithFeeDest is the multiLocation of the asset that is intended to be used to pay for fees in the dest chain
 				paysWithFeeDest:
 					'{"parents": "1", "interior": {"X3": [{"Parachain": "1000"}, {"PalletInstance": "50"}, {"GeneralIndex": "1984"}]}}',