+++ examples/relayToPara.ts
@@ -9,7 +9,7 @@ import { GREEN, PURPLE, RESET } from './colors';
 
 /**
  * In this example we are creating a call to send 1 KSM from a Kusama (Relay Chain) account
- * to a Moonriver (Parachain) account, where the `xcmVersion` is set to 2 and no `weightLimit` is provided declaring that
+ * to a Moonriver (Parachain) account, where the `xcmVersion` is set to safeXcmVersion and no `weightLimit` is provided declaring that
  * the tx will allow unlimited weight for fees.
  *
  * NOTE: To specify the amount of weight for the tx to use provide a `weightLimit` option containing desired values for `refTime` and `proofSize`.
@@ -26,7 +26,7 @@ const main = async () => {
 			['1000000000000'],
 			{
 				format: 'call',
-				xcmVersion: 2,
+				xcmVersion: safeXcmVersion,
 			},
 		);