+++ src/createXcmTypes/RelayToBridge.spec.ts
@@ -2,15 +2,15 @@
 
 import { Registry } from '../registry';
 import { adjustedMockSystemApiV1011000 } from '../testHelpers/adjustedMockSystemApiV1011000';
-import { SystemToBridge } from './SystemToBridge';
+import { RelayToBridge } from './RelayToBridge';
 
-describe('SystemToBridge', () => {
-	const registry = new Registry('asset-hub-rococo', {});
+describe('RelayToBridge', () => {
+	const registry = new Registry('rococo', {});
 	const isForeignAssetsTransfer = true;
 	const isLiquidTokenTransfer = false;
 	describe('Beneficiary', () => {
 		it('Should work for V3', () => {
-			const beneficiary = SystemToBridge.createBeneficiary(
+			const beneficiary = RelayToBridge.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				3,
 			);
@@ -31,7 +31,7 @@ describe('SystemToBridge', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3 for an Ethereum Address', () => {
-			const beneficiary = SystemToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 3);
+			const beneficiary = RelayToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 3);
 
 			const expectedRes = {
 				V3: {
@@ -49,7 +49,7 @@ describe('SystemToBridge', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const beneficiary = SystemToBridge.createBeneficiary(
+			const beneficiary = RelayToBridge.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				4,
 			);
@@ -72,7 +72,7 @@ describe('SystemToBridge', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4 for an Ethereum Address', () => {
-			const beneficiary = SystemToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 4);
+			const beneficiary = RelayToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 4);
 
 			const expectedRes = {
 				V4: {
@@ -95,11 +95,11 @@ describe('SystemToBridge', () => {
 	describe('Destination', () => {
 		it('Should work for V3', () => {
 			const destId = `{"parents":"2","interior":{"X1":{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}}}}`;
-			const destination = SystemToBridge.createDest(destId, 3);
+			const destination = RelayToBridge.createDest(destId, 3);
 
 			const expectedRes = {
 				V3: {
-					parents: 2,
+					parents: 1,
 					interior: {
 						X1: {
 							GlobalConsensus: {
@@ -117,11 +117,11 @@ describe('SystemToBridge', () => {
 		it('Should work for V4', () => {
 			const destId = `{"parents":"2","interior":{"X2":[{"GlobalConsensus":"Kusama"},{"Parachain":"1000"}]}}`;
 
-			const destination = SystemToBridge.createDest(destId, 4);
+			const destination = RelayToBridge.createDest(destId, 4);
 
 			const expectedRes = {
 				V4: {
-					parents: 2,
+					parents: 1,
 					interior: {
 						X2: [
 							{
@@ -140,11 +140,11 @@ describe('SystemToBridge', () => {
 	});
 	describe('Assets', () => {
 		it('Should work for V3', async () => {
-			const assets = await SystemToBridge.createAssets(
+			const assets = await RelayToBridge.createAssets(
 				['10000000000'],
 				3,
-				'asset-hub-westend',
-				[`{"parents":"2","interior":{"X1":{"GlobalConsensus":"Rococo"}}}`],
+				'rococo',
+				[`{"parents":"0","interior":{"Here":""}}`],
 				{
 					registry,
 					isForeignAssetsTransfer,
@@ -158,9 +158,9 @@ describe('SystemToBridge', () => {
 					{
 						id: {
 							Concrete: {
-								parents: '2',
+								parents: 0,
 								interior: {
-									X1: { GlobalConsensus: 'Rococo' },
+									Here: '',
 								},
 							},
 						},
@@ -174,13 +174,11 @@ describe('SystemToBridge', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', async () => {
-			const assets = await SystemToBridge.createAssets(
+			const assets = await RelayToBridge.createAssets(
 				['10000000000'],
 				4,
-				'asset-hub-westend',
-				[
-					`{"parents":"2","interior":{"X2":[{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}},{"AccountKey20":{"network":null,"key":"0xfff9976782d46cc05630d1f6ebab18b2324d6b14"}}]}}`,
-				],
+				'rococo',
+				[`{"parents":"0","interior":{"Here":""}}`],
 				{
 					registry,
 					isForeignAssetsTransfer,
@@ -193,23 +191,9 @@ describe('SystemToBridge', () => {
 				V4: [
 					{
 						id: {
-							parents: '2',
+							parents: 0,
 							interior: {
-								X2: [
-									{
-										GlobalConsensus: {
-											Ethereum: {
-												chainId: '11155111',
-											},
-										},
-									},
-									{
-										AccountKey20: {
-											network: null,
-											key: '0xfff9976782d46cc05630d1f6ebab18b2324d6b14',
-										},
-									},
-								],
+								Here: '',
 							},
 						},
 						fun: {